+++ .wings/bootstrap.js
@@ -1,7 +1,38 @@
-const Wings = require('@wingscms/sdk').default;
 const fs = require('fs');
 const path = require('path');
+const { GraphQLClient } = require('graphql-request');
 const { getIntrospectionQuery } = require('graphql');
+
+const WINGS_APP_KEY = 'z4SS8rqLDrNWMChzslf3IBuWJG1ldbSy';
+
+class Wings {
+  constructor({ endpoint = 'https://api.wings.dev', appKey, project, domain }) {
+    this.endpoint = endpoint;
+    this.appKey = appKey;
+    this.project = project;
+    this.domain = domain;
+
+    const authHeaders = domain
+      ? {
+          Authorization: `Bearer ${WINGS_APP_KEY}`,
+          'X-Wings-App-Domain': domain,
+        }
+      : {
+          Authorization: `Bearer ${appKey}`,
+          'X-Wings-Project': project,
+        };
+
+    this.client = new GraphQLClient(endpoint, {
+      headers: {
+        ...authHeaders,
+      },
+    });
+  }
+
+  async query(...args) {
+    return this.client.request(...args);
+  }
+}
 
 const schemaPath = path.join(__dirname, 'data', 'introspectionResult.json');