+++ packages/components/src/theme/Theme.js
@@ -90,6 +90,14 @@ export default class Theme {
     return this.variables.baseTabletFontSize || this.baseFontSize;
   }
 
+  get appBarBackgroundColor() {
+    return this.variables.appBarBackgroundColor || this.elementBackgroundColor;
+  }
+
+  get appBarHeight() {
+    return this.variables.appBarHeight || this.largeSpacing;
+  }
+
   get burgerColor() {
     return this.variables.burgerColor || this.iconColor;
   }
@@ -120,6 +128,19 @@ export default class Theme {
 
   get extraSmallSpacing() {
     return this.variables.extraSmallSpacing || this.calc(this.mediumSpacing, ms => ms / 4);
+  }
+
+  get footerBackgroundColor() {
+    return this.variables.footerBackgroundColor || this.primaryColor;
+  }
+
+  get footerTextColor() {
+    return (
+      this.variables.footerTextColor ||
+      this.contrastColor({
+        backgroundColor: this.footerBackgroundColor,
+      })
+    );
   }
 
   get formBackgroundColor() {
@@ -176,6 +197,10 @@ export default class Theme {
     return this.variables.linkColor || this.primaryColor;
   }
 
+  get linkSecondaryColor() {
+    return this.variables.linkSecondaryColor || this.secondaryColor;
+  }
+
   get navigationIconColor() {
     return this.variables.navigationIconColor || this.iconColor;
   }
@@ -203,6 +228,14 @@ export default class Theme {
 
   get navigationMenuBackgroundColor() {
     return this.variables.navigationMenuBackgroundColor || this.backgroundColor;
+  }
+
+  get scrollBarColor() {
+    return this.variables.scrollBarColor || this.primaryColor;
+  }
+
+  get scrollBarBackgroundColor() {
+    return this.variables.scrollBarColor || this.elementBackgroundColor;
   }
 
   get shareButtonBackgroundColor() {