+++ packages/components/src/theme/defaultTheme.js
@@ -20,6 +20,7 @@ export default {
   mediumSpacing: '40px',
   // misc
   contrastLuminanceThreshold: 60,
+  linkStyle: 'lineGrow',
   shadow: '0 0 40px 0 rgba(0, 0, 0, 0.05)',
   // typography
   baseFontSize: '16px',