+++ packages/react/src/components/Article.js
@@ -2,6 +2,6 @@ import React from 'react';
 import { Header } from '@wingscms/components';
 import Entry from './Entry';
 
-export default props => {
+export default function Article(props) {
   return <Entry {...props} headerType={Header.Type.COVER} />;
-};
+}