+++ packages/react/src/components/Campaign/Campaign.js
@@ -146,7 +146,7 @@ const CounterContainer = styled.div`
   }
 `;
 
-export default ({
+export default function Campaign({
   id,
   resourceType,
   node: _node = {},
@@ -157,7 +157,7 @@ export default ({
   signatureCount: _signatureCount,
   signatureGoal: _signatureGoal,
   ...props
-}) => {
+}) {
   const intl = useIntl();
   const { width } = useDimensions();
   const campaignContainerRef = useRef(null);
@@ -295,4 +295,4 @@ export default ({
     </>
   );
   return wrapElement(element, node);
-};
+}