+++ packages/react/src/components/MobiledocRenderer/markups.js
@@ -1,0 +1,3 @@
+import { Link } from '@wingscms/components';
+
+export default [{ name: 'a', render: Link }];